# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Viv Scanner - Firmware Security Analysis Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/vulnerabilities/{id}:
    get:
      operationId: getVulnerabilities
      tags: [security]
      summary: Get vulnerabilities for firmware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vulnerability list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vulnerability"
  /security/score/{id}:
    get:
      operationId: getSecurityScore
      tags: [security]
      summary: Get security score for firmware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Security score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityScore"
  /security/secrets/{id}:
    get:
      operationId: getHardcodedSecrets
      tags: [security]
      summary: Get hardcoded secrets found in firmware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Secrets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HardcodedSecret"
  /security/dangerous-functions/{id}:
    get:
      operationId: getDangerousFunctions
      tags: [security]
      summary: Get dangerous function calls found in firmware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dangerous function list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DangerousFunction"
...
components:
  schemas:
...
    Vulnerability:
      type: object
      required: [id, firmwareId, type, severity, affectedFile, description, recommendation]
      properties:
        id:
          type: integer
        firmwareId:
          type: integer
        type:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        affectedFile:
          type: string
        description:
          type: string
        recommendation:
          type: string
        cvssScore:
          type: ["number", "null"]
        discoveredAt:
          type: string
    SecurityScore:
      type: object
      required: [firmwareId, overallScore, riskLevel, criticalCount, highCount, mediumCount, lowCount]
      properties:
        firmwareId:
          type: integer
        overallScore:
          type: integer
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        hardcodedSecretsCount:
          type: integer
        dangerousFunctionsCount:
          type: integer
    HardcodedSecret:
      type: object
      required: [id, firmwareId, type, value, file, line]
      properties:
        id:
          type: integer
        firmwareId:
          type: integer
        type:
          type: string
          enum: [password, api_key, token, private_key, certificate]
        value:
          type: string
        file:
          type: string
        line:
          type: integer
        severity:
          type: string
          enum: [critical, high, medium, low]
    DangerousFunction:
      type: object
      required: [id, firmwareId, name, file, line, risk]
      properties:
        id:
          type: integer
        firmwareId:
          type: integer
        name:
          type: string
        file:
          type: string
        line:
          type: integer
        risk:
          type: string
          enum: [critical, high, medium]
        description:
          type: string
...